Make sure that file name and defined class on the file matches.
By following the Ruby convention that most Ruby developers are used to,
makes it easy to find the files and classes on our codebase.

I excluded certain files to avoid having to do a big rename.
For example to have a class name be OpenTelemtry we would
need the filename or directory be open_telemetry.
Those changes makes looking at the git history hard, and the might not
provide a lot of value.

Also, there are certain names in the industry that will not match the ruby
conventition such as MongoDB.