Introduce `sample_labels` struct to carry labels passed to libdatadog
This is a refactor in preparation for introducing the new `state` label.

The `state` label is different from other labels in that it will be set
after looking at the thread stack. Thus, I needed a way to get more
information about labels from the `ThreadContext` collector to the
`Stack` collector, and thus the `sample_labels` struct was born.