[PROF-7440] Add approximate thread state categorization for timeline
**What does this PR do?**

This PR adds a new feature to the Ruby profiler: thread state
categorization.

When taking a cpu/wall-time sample, the profiler will peek at the
top of the stack, and then add a `state` label to the sample that
categorizes what the thread was doing.

**Motivation:**

By categorizing what threads are doing when they're not on the cpu,
we can synthesize nice events for timeline, e.g. paint a big
rectangle when a thread was blocked on some mutex, or was waiting for
the network.

**Additional Notes:**

Because we're matching only on "name of method at the top of the stack"
+ "is this method a native method" this approach is approximate, and
we hope to replace it with a more accurate mechanism soon.

For now, we will only expose this information as part of the timeline
feature.

A few tests needed to be refactored, because they assumed aggregation
on the samples that they took, and the new label makes samples not
aggregate some times.

I considered if the tests should be changed, or the feature, but
concluded that the tests were, in a sense, wrong for assuming
aggregation, because in the future, we'll have timeline on by default
and in such a future aggregation never happens.

**How to test the change?**

This change includes code coverage. You can also test it by enabling
the timeline feature, and observing that the new states show up
on Ruby timeline profiles (you'll need a frontend feature flag to
see timeline profiles).