[PROF-7307] Enable allocation counting feature by default for some Ruby 3 versions
**What does this PR do?**

This PR enables the Profiler's `allocation_counting_enabled` feature
by default for some Ruby 3 versions.

(This feature depends on the profiler itself being enabled).

TL;DR we were already enabling this feature by default on Ruby 2, but
did not do it on Ruby 3 because of a VM bug
( https://bugs.ruby-lang.org/issues/18464 ).

Now that this VM bug has been fixed for some Ruby versions
(3.1.4, 3.2.3, 3.3.0), we can enable it by default again.

**Motivation:**

The allocation counting feature is required to enable other
features in the future (including allocation profiling) and thus
we want to allow as many customers as possible to have it.

**Additional Notes:**

There's a second annoying VM bug that can also break
`allocation_counting_enabled` (https://bugs.ruby-lang.org/issues/19112)
but since that bug doesn't cause a VM crash and only makes it so that
data collection stops for this feature, I think we can live with it
for now.

I'm also reaching out through some contacts to see if we can get the
other bug fixed soon (and possibly backported).

**How to test the change?**

Change includes test coverage.