Profiling: Add note about Ruby 3.2 bug
**What does this PR do?**

This PR adds a reference to https://bugs.ruby-lang.org/issues/19482 /
https://github.com/ruby/ruby/pull/7464 to the reasons why
`allocation_counting_enabled` should not be enabled on Ruby 3.2.0 to
3.2.2.

No code needs to be changed since we were already excluding those
Ruby versions due to a different bug.

**Motivation:**

I like having this in the code as it may be relevant in the future
(e.g. if the issue reappears).

**Additional Notes:**

N/A

**How to test the change?**

No code is changed!