Try to fix flaky profiler spec
**What does this PR do?**

This test attempts a fix to the flaky profiler tests that were added in
 #3162 (and that bit us in #3179).

I'm not able to reproduce the issue locally, but I suspect the issue
is that the spec as written was racy:

1. We start a background thread and synchronize with it via a queue
2. The thing we assert on happens AFTER the synchronization (e.g.
   some method called after the queue is used to signal)

Thus, I suspect the issue is that once in a while, Ruby switches
threads at just the wrong time: between 1 and 2, and thus our
assertion fails.

Instead, I've changed the specs  to have a `wait_for` step that waits
for the stack trace of the thread to be what we previously only assumed
it to be, and only then do we run the other assertions.

I call this fix "an attempt" since I wasn't able to reproduce the
issue, BUT overall it's harmless: either I got it right and this
fixes it for got, or I got it wrong, and we'll keep seeing the
flaky specs, and I'll figure out something else to try.

**Motivation:**

Profiler always aims to have no flaky tests.

**Additional Notes:**

N/A

**How to test the change?**

Validate that CI is still green, and stays that way ;)

Fixes datadog/ruby-guild#53