Shut down profiler if any components failed
**What does this PR do?**

This PR changes the profiler `Collectors::CpuAndWallTimeWorker` and
`Scheduler` components to signal the top `Profiler` instance when they
failed, so the `Profiler` can stop the other component as well.

**Motivation:**

While never reported by a customer, it ocurred to me that right now if
one of the component fails, the other will keep working in a weird
state.

Of particular concern, if the scheduler failed, the `StackRecorder`
would trigger a memory leak: it would start accumulating samples and
never flush them.

This PR changes the profiler so that if one component stops with
a failure, the other will do so as well.

**Additional Notes:**

I'm not especially happy with how this change turned out.

1. The error handling in the scheduler somewhat duplicates similar
   logic in the `Core::Workers::Async` helper, which it uses.

   I wanted to avoid changing the shared helper just for this
   use-case, which I managed, but at the cost of a bit of
   duplication (including a duplicated log message, sigh).

2. I don't like the `on_failure_proc: nil`, but adding the
   argument would mean changing A LOT of tests, which is doable
   but has its own ugliness as well.

   I've considered passing the `on_failure_proc` via the constructor,
   but it's not entirely straightforward: the `Profiler` instance is
   created after the two others, so we'd need a way to break the loop.

In the end, I decided to move forward with something, as I think it's
important to plug this gap.

**How to test the change?**

This change includes test coverage.