Upgrade steep typechecker to version 1.6.0
**What does this PR do?**

Upgrade the steep typechecker to version 1.6.0.

**Motivation:**

The steep typechecker is under very active development so there's a
number of improvements to be picked up, see
https://github.com/soutaro/steep/wiki/Release-Note-1.6 .

In particular, I'm very interested in improvements to error messages,
since I still think that's a particular weak point for steep.

**Additional Notes:**

N/A

**How to test the change?**

Validate that CI is still green.