Mark ddtrace threads as fork-safe
**What does this PR do?**

This PR applies to all relevant ddtrace threads a pattern pioneered by
rails: set the `:fork_safe` thread variable to indicate that it's OK
if ddtrace threads are started before a Ruby app calls fork.

**Motivation:**

This PR fixes the issue reported in #3203. The puma webserver in
clustering mode has a check where it will warn when it detects that
threads were started before the webserver has fork'd the child
processes, see:

https://github.com/puma/puma/blob/32e011ab9e029c757823efb068358ed255fb7ef4/lib/puma/cluster.rb#L353-L359

This is not a problem for ddtrace (as discussed in #3203), but the
warning is annoying.

While looking into the issue again today, I spotted that actually
there's a way to tell puma that a thread is fine and is it's not a
problem if it's started before the webserver has fork'd: puma
checks if the thread has a `:fork_safe` thread-variable set to `true`.

**Additional Notes:**

We have a bunch of places in ddtrace where we create threads...
I briefly played with the idea of creating a superclass of all
ddtrace Threads, but decided to not get into a bigger refactoring
for such a small issue. Maybe next time...?

**How to test the change?**

Test coverage included. Furthermore, you can try running
a Ruby webapp in puma before/after the change to confirm the
warning is gone.

Fixes #3203