Avoid using `-Wunused-parameter` on Ruby 3.3
As documented in `extconf.rb`, this flag can cause compilation failures
because mkmf picks it up and uses it when calling `have_header`.

For now I decided to disable it on 3.3, for fear it'll happen in
3.3.0-stable as well. We can revisit this later, and anyway for our
own we get a lot of cover from the fact that all other Ruby versions
still compile with this option enabled.