Revert "Disable Ruby 3.3 integration due to protobuf compilation error"
**What does this PR do?**

This reverts commit f9ebce90b92fbb0fec14f0865504610b31b51a04.

**Motivation:**

In #3211 we disabled our CI integration testing on Ruby 3.3 because
one of the dependencies (protobuf) had an incompatibility at the
time (see protocolbuffers/protobuf#14509 for details).

Upstream has since fixed the issue, so we're good to go.

**Additional Notes:**

N/A

**How to test the change?**

Check that CircleCI is now also running the
`build_and_test_integration-3.3` test.