Call Devise RegistrationsController block
When a Devise user calls create with a block, that block was swallowed
by the block we use to perform our resource information extraction for
registration events.

Ensure the original block is called. We call it after our analysis in
the event we should block for some reason, thus protecting the
additional user code in their block.

JIRA:
- [SCRS-704](https://datadoghq.atlassian.net/browse/SCRS-704)
- [APPSEC-12115](https://datadoghq.atlassian.net/browse/APPSEC-12115)
- [APPSEC-12770](https://datadoghq.atlassian.net/browse/APPSEC-12770)