[NO-TICKET] Minor: Clean up a few classes kept in global variables
**What does this PR do?**

This PR gets rid of two references to classes that were being kept as
global variables in the profiling C extension.

* The one in `stack_recorder.c` wasn't actually needed
* The one in `http_transport.c` was only needed for an upcall to Ruby
  code in a rare error handling case, so I've replaced it with an
  actual lookup, so we don't need to keep the reference

**Motivation:**

Having as little global state as possible seems like a good idea, and
we don't need to care about telling or not telling the Ruby GC about
these.

**Additional Notes:**

N/A

**How to test the change?**

The change in `http_transport.c` already has code coverage -- I checked
this by quickly writing the wrong code and validating that indeed
this code was exercised and some of the tests failed as expected.