[NO-TICKET] Remove unused parameter warning workaround for Ruby 3.3
**What does this PR do?**

This PR removes the workaround we add to not use the
`-Wunused-parameter` warning on Ruby 3.3.

This means we'll have those warnings when compiling the profiler
native extension on that Ruby version.

The issue was that some of the internal Ruby headers in Ruby 3.3
have unused parameters. If we applied the warning before checking
these internal headers, the header detection check (`have_header`)
would fail not because the headers were missing, but because they
have warnings.

(Why do they fail due to warnings? Because by default `have_header`
tries to compile a tiny app that includes the header AND tries
to compile that app with `-Werror`, turning any warning into an
error -- thus making the combination fail)

To fix this issue, we need to enable these warnings after all the other
checks have passed, and after nothing else will be checked with
`-Werror`.

**Motivation:**

This fix enables us to have this useful warning enabled when
compiling the profiler for Ruby 3.3 (and probably for future
versions of Ruby).

**Additional Notes:**

This "internal headers with warnings cause us not to be able to
use -Wunused-parameter" also happens with some older Ruby versions,
on some of the public headers.

On those versions `add_compiler_flag '-Wunused-parameter'` will
try to add the flag and fail (same reason as my note above:
headers have warnings + check uses `-Werror`).

This PR was born out of an attempt of fixing that problem as
well, but I didn't like the fix, because while I can fix
`extconf.rb` (`try_cflags` can recieve an option to disable
`-Werror`), the next problem is thatt when compiling our own
files, a torrent of warnings shows up from those internal
headers. And... getting rid of those warnings so we can
see anything else is... too ugly. So for now I've extracted
only the relevant part of my fix.

**How to test the change?**

This whole thing only really affects development, so if CI is
still green, things are still fine.

Additionally, you can remove a `DDTRACE_UNUSED` qualifier in
some of the functions and observe that we now get warnings for
unused parameters on the Ruby 3.3 image we use for CI.