Garbage Collection samples no longer collect stacks
When this feature was added, we weren't yet working an allocation
profiling, so having stacks would hint at where an allocation was that
triggered the GC.

This... is basically a very crappy allocation profiler, and as
an anti-bonus it can also get quite expensive.

As part of making this feature shape up for us to have it on by
default, let's no longer collect stacks. As an alternative, you
can get a similar context either from allocation profiling OR by
looking at thread timeline for what the thread was doing around the
time the GC happened.