Remove record_buffer/extra_frames_in_record_buffer abstraction
These abstractions added some complexity to the stack collector, and
were only needed to allow us to put extra placeholder frames on top
of a true collected stack.

Since this was only being used by the GC placeholder feature, and
we've changed that to not collect a stack anymore, we can clean up
this mechanism and simplify the code.