Allow the dynamic sampling rate overhead target to be set
The profiler has a sophisticated dynamic sampling rate mechanism to tune
the sample collection rate in order to target a desired level of
wall-clock overhead time. Unfortunately, the default value of 2% is too
much for our application; we find a noticeable and unfortunately not
acceptable impact on our application's p95/p99 response times.

This commit adds a configuration option
dynamic_sampling_rate_overhead_target_percentage which allows this to be
controlled. The tradeoff for reducing the overhead target is that the
profiles will have fewer samples in them and hence be less accurate.