[PROF-8667] Heap Profiling - Part 3 - Snapshot system (#3328)
This commit follows bb96f8f87e30e920cf8c53d4ddba9683b82b5ae5 and reduces the conflict of serialization and sampling which was forcing us to potentially drop some allocation samples on the floor. It also removes the need for heap recorder to have to use a mutex by switching to a iterate-over-snapshot strategy.