[NO-TICKET] Import upstream `rb_profile_frames` fix
**What does this PR do?**

This PR imports an upstream fix to `rb_profile_frames` from
https://github.com/ruby/ruby/pull/9310 into our own custom copy of
that function.

**Motivation:**

As explained in the comment, the situation being handled does not yet
happen with dd-trace-rb in practice since we don't support profiling
with MN threads enabled, but since we want to try supporting it at
some point, I've gone ahead and imported the fix.

**Additional Notes:**

N/A

**How to test the change?**

This is a bit awkward to test, so didn't add any coverage.