Bump debase-ruby_core_source dependency to 3.3.0
**What does this PR do?**

This PR bumps our dependency on `debase-ruby_core_source` to version
3.3.0. This version includes the headers for the recently-released
Ruby 3.3.0; up until now we were supporting that Ruby version using
a copy of the headers from the preview release.

**Motivation:**

Use up-to-date headers for profiling on Ruby 3.3.0.

**Additional Notes:**

I've updated the appraisal gemfiles as well.

**How to test the change?**

Validate that CI is green.