Use `rb_postponed_job_preregister`/`rb_postponed_job_trigger` on Ruby 3.3
**What does this PR do?**

This PR changes the profiler to use the new
`rb_postponed_job_preregister`/`rb_postponed_job_trigger` API instead
of `rb_postponed_job_register_one` in Ruby 3.3.

The old API is deprecated.

**Motivation:**

Usage of the old API was causing CI to fail, since we turn every
warning into an error in CI, and Ruby 3.3 includes a nice warning
when the deprecated API gets used.

**Additional Notes:**

See also https://bugs.ruby-lang.org/issues/19991 and
https://github.com/ruby/ruby/pull/8949 if you're curious on why
this new API was introduced.

**How to test the change?**

The code path for calling `sample_from_postponed_job` is already
covered by existing tests.

I discovered while working on this that the code path for calling
`after_gc_from_postponed_job` from `on_gc_event` is missing
coverage since we exercise that behavior in a different way in the
specs.

I've taken a note to circle back on this, but I didn't want to
delay the PR since CI is broken in master currently.