[NO-TICKET] Remove leftover helper file for profiling 0.x to 1.x migration
**What does this PR do?**

This PR removes the `ddtrace/profiling/preload` file. As the comment
attached to it said, this file was used to direct users to the correct
new file in the 0.x to 1.x migration.

Let's remove it for 2.0.

**Motivation:**

Empty out the `lib/ddtrace` folder.

**Additional Notes:**

N/A

**How to test the change?**

Validate that `ddtrace/profiling/preload` no longer exists? ;)