Extract `convert_tags` and `endpoint_from` from `HttpTransport` to libdatadog helpers
These are going to be needed also for the crash tracker code.

While doing this extraction, I've gone ahead and made the failure
logging on `convert_tags` use the ddtrace logger directly, rather than
having to rely on an extra method to do this conversion.

Since these methods are covered by the tests in
`http_transport_spec.rb`, I've chosen not to expose them in other ways
for testing.