Improve performance of gathering the ClassCount metric
**What does this PR do?**

This PR tweaks the `ClassCount.available?` method to cache its result.

The result is expected to be always the same for a given Ruby VM, so
we don't need to keep checking again and again.

**Motivation:**

While looking at our GitLab test deployments, I noticed that on this
big app, calling `ObjectSpace.count_objects` does have some cost
(because this method goes through the whole Ruby heap).

I then noticed we _always_ call it twice in a row -- we call
`available?` and then call `value`:

![image](https://github.com/DataDog/dd-trace-rb/assets/2785847/a6f9fc20-058a-4018-babc-95e93ccbe06d)

It's trivial to avoid this, so I decided to open this tiny PR.

**Additional Notes:**

N/A

**How to test the change?**

Existing test coverage is enough :)