[NO-TICKET] Fix sinatra integration apps broken by missing `rackup`
**What does this PR do?**

This PR adds the `rackup` as a dependency of the sinatra integration
apps.

This fixes our sinatra-based integration apps being broken in CI for
master.

The `rackup` command is no longer included in modern versions
of `rack`, and has

**Motivation:**

The `rackup` command is no longer included in modern versions of
`rack`, and has been extracted into a separate gem.

The release of `sinatra` 4 moved to one such modern version of
`rack`, which broke our integration apps.

**Additional Notes:**

Sigh, this is the second "CI-broken-in-master" in 7 days :(

**How to test the change?**

Validate that CI is green!