[NO-TICKET] Pin sinatra integration apps to sinatra 2
**What does this PR do?**

This PR pins our sinatra integration apps to make sure they use
sinatra 2. They were otherwise picking up sinatra 3 and 4.

It also tweaks a few of the gems to avoid future troubles when
trying to run this app with newer sinatra/rack versions.

**Motivation:**

This PR is an alternative to #3390. Our sinatra apps are broken in CI
as they picked up sinatra 4 and started breaking due to a minor
incompatibility (`rack` 3 no longer includes `rackup` by default).

We could've opted to use newer sinatra versions for these apps,
but our current pattern with rails is to have a different app for
different versions, so for now let's keep with it.

**Additional Notes:**

N/A

**How to test the change?**

Validate that CI is green.